+++ README.md
@@ -2,10 +2,36 @@
 
 ## Version 2
 
-A personal project to explore Javascript
+last updated : november 2023
+
+A word game to explore and learn Javascript
 
 Live version at https://word-game-2023.netlify.app/
 
-- updated script for more efficient game refresh.
+# New features
 
-- stat show on start up
+- Start game button now present on first load / page refresh
+
+- Game can be started from inside stats panel
+
+# Bug fixes
+
+- New games do not refresh the web page. Only one call to web server on first load
+
+- Fixed color flashing on first page load when in dark mode
+
+- removed dependency for animate.css external library
+
+- reseting stats / clear stats now show instantly in stats panel.
+
+- Game can be restarted mid way through another game without letters repeating from previous game.
+
+# Future improvments to come
+
+- Add 2 step to reseting stats incase button is accidentaly clicked on.
+
+- timed mode
+
+- more animated themes
+
+- game complete animations
